Mix in a MacOS test and different Linux distributions in Travis
Due to the similarity between Linux and MacOS, there's no need to have
every Python version on MacOS tested. Instead, this commit picks one
Python version and run it on MacOS, in case of some surprise. This
reduces the cluttering of tests.

Also, this commit varies linux dist environment a bit to make the CI
more solid.