Pin the versions of the packages that are used in tests and add dependabot
We should pin the versions of all Python packages that we are using
solely for testing and doc
generating for a stable test and doc env (e.g., future incompatibility,
regression, etc.). We want
to pin these because, in this project, We use these packages soly for
deployment of our development
environment (i.e., running tests and generating docs) and we want stable
packages that are used by
us for these purposes. We should let some dependency bot to verify that
bumping the versions won't
break anything before we actually upgrade any of these dependencies.

We should not pin the actual dependencies of PyDAX (as specified in
`setup.py`), because PyDAX is an
intermediate software layer -- those should be pinned only by the actual
deployed application that
depends on PyDAX. We should only code the info of supported versions of
these dependencies. If there
is some regression or incompatibilities in the latest versions of our
dependencies, we should either
work around them, or update `setup.py` to avoid depending on those
versions.