Add basic tests for HTTPS
Our test infrastructure should now be able to support HTTPS tests locally.

`test_custom_schema` now should be able to test over a https link
properly.

Also the https fixtures are a bit similar to the http fixtures, but they
are also different enough that reusing their code seems to be not that
easy.

Note that, one reason that we don't use openssl to generate the key and
certificate here is that it involves a lot of tweaks (and many more
commands) to generate a one-off self-signed certificate with multiple
Subjective Alternative Names (SANs) specified (at least when we first
tried at the end of 2020).