pydax_initialization should localize the download URLs
This change localizes the URLs in the initialization so that the
reliance on remote hosts is minimized.

After this change, all tests can run without Internet connections after
the first run (which downloads relevant datasets locally). This can
potentially speed up tests if the Internet connections are not very
fast.