Add README badges (#114)
The version badge and PyPI badge are placeholders (we haven't published
on pypi yet). The Docs CI badge should be a sufficient replacement for
readthedoc badge (we will be moving the document to somewhere else in
the future). Coverage badge isn't here because some admin needs to
install the coverall service.